objdump: construct Object Header Table entries
Each object embedded in a multi-object ELF file, including the original
ELF binary's object onto which additional objects are transplanted, has
an object header providing information required by a dynamic linker to
link and relocate the object.

The table is stored in a section called ".object" of a new segment with
the PT_OBJECT type. The original object is placed as the first in the
table.

(cherry picked from commit 1c1b1788fb75bbafd94e587eba67bcc7565b9cb6)