posix shm: rework capability support
In order to preserve provenance and prevent colluding proceses from
hiding capabilities from the revoker, we must not share capabilities
across address spaces.  However, doing so is often useful and some
existing code makes use of of this functionality in relatively safe
ways.

As a compromise, allow sharing, but only if the programmer makes their
intent clear by making the following changes:
 - Require that PROT_CAP be requested explicitly in mmap to allow load
   or store of capabilities.
 - By default, attach each share memory object to the first address
   space (vmspace) that opens it and allow only that address space to
   map the object with PROT_CAP.  On fork, remove capability permissions
   from the mapping.
 - Add an O_SHARECAP flag for shm_open2 (aka shm_open and memfd_create)
   which allows sharing capabilities across address spaces, overriding
   the default.