posix shm: revoke local objects
During revocation, map all local shm objects into the address space to
ensure that they are scanned even when they are not (fully) mapped.
Mappings created with O_SHARECAP aren't mapped as revocation on them
is unreliable by design.

I had originally envisioned a model were we maintained a mapping of
only pages not currently mapped in the address space, but the required
bookkeeping seems quite complex and there isn't obvious machinery to
handle it.