c18n: Improve detection of interrupted compartment
Previously, the interrupted compartment is identified by checking
whether its untrusted stack is a subset of the canonical copy of that
stack. However, this subset relation should actually be inverted, that
is, the untrusted stack of the interrupted compartment should be a
superset of the canonical stack. Otherwise, a compartment which holds a
portion of another compartment's stack would be able to impersonate it.
(The SW_VMEM permission on the canonical stack has to be stripped before
the comparison.)

In addition, a bounds-check was missing for the tail-call scenario,
which might result in reading from the stack table using an invalid
index.

Also, the tail-call scenario cannot be ruled-out even when the
presumptive callee's stack is not allocated---we might be interrupted
when a tail-call is taking place _and_ the callee's stack is
unallocated.